+++ bl2/ext/mcuboot/bootutil/src/bootutil_priv.h
@@ -201,7 +201,7 @@ struct boot_loader_state {
     } scratch;
 
     uint8_t swap_type[BOOT_IMAGE_NUMBER];
-    uint8_t write_sz;
+    uint32_t write_sz;
 };
 
 uint32_t boot_secure_memequal(const void *s1, const void *s2, size_t n);
@@ -209,7 +209,7 @@ int bootutil_verify_sig(uint8_t *hash, uint32_t hlen, uint8_t *sig,
                         size_t slen, uint8_t key_id);
 
 int boot_magic_compatible_check(uint8_t tbl_val, uint8_t val);
-uint32_t boot_trailer_sz(uint8_t min_write_sz);
+uint32_t boot_trailer_sz(uint32_t min_write_sz);
 int boot_status_entries(const struct flash_area *fap);
 uint32_t boot_status_off(const struct flash_area *fap);
 uint32_t boot_swap_info_off(const struct flash_area *fap);
